+++ cdk.context.json
@@ -1,0 +1,18 @@
+{
+  "hosted-zone:account=834212829299:domainName=kovvuri.people.aws.dev:region=834212829299": {
+    "Id": "/hostedzone/Z00657381W5TZZFS20M36",
+    "Name": "kovvuri.people.aws.dev."
+  },
+  "availability-zones:account=834212829299:region=us-east-1": [
+    "us-east-1a",
+    "us-east-1b",
+    "us-east-1c",
+    "us-east-1d",
+    "us-east-1e",
+    "us-east-1f"
+  ],
+  "hosted-zone:account=834212829299:domainName=kovvuri.people.aws.dev:region=us-east-1": {
+    "Id": "/hostedzone/Z00657381W5TZZFS20M36",
+    "Name": "kovvuri.people.aws.dev."
+  }
+}
