--- src/backend/flask/Dockerfile
@@ -1,17 +1,0 @@
-FROM python:3.9-slim
-
-WORKDIR /app
-
-COPY requirements.txt /app
-
-RUN pip install --no-cache-dir -r requirements.txt
-
-COPY *.py /app 
-
-COPY output/ /app 
-
-EXPOSE 5000
-
-ENV FLASK_APP=app.py
-
-CMD ["flask", "run", "--host=0.0.0.0", "--port=5000"]
