--- src/backend/flask/requirements.txt
@@ -1,33 +1,0 @@
-asgiref==3.7.2
-blinker==1.6.2
-certifi==2024.7.4
-charset-normalizer==3.1.0
-click==8.1.3
-Django==4.2.14
-Flask==2.3.2
-idna==3.7
-image==1.5.33
-itsdangerous==2.1.2
-Jinja2==3.1.4
-jmespath==1.0.1
-MarkupSafe==2.1.2
-python-dateutil==2.8.2
-requests==2.32.2
-s3transfer==0.6.1
-six==1.16.0
-sqlparse==0.5.0
-urllib3==1.26.19
-Werkzeug==3.0.3
-faiss-cpu==1.7.4
-pypdf==3.17.0
-pypdf2==3.0.1
-langchain==0.1.11
-anthropic==0.2.10
-tika==2.6.0
-bs4==0.0.1
-fpdf==1.7.2
-selenium==4.11.2
-flask-cors==4.0.1
-opensearch-py==2.3.0
-boto3
-langchain-community
