--- src/frontend/amplify/.config/project-config.json
@@ -1,17 +1,0 @@
-{
-  "projectName": "bedrock",
-  "version": "3.1",
-  "frontend": "javascript",
-  "javascript": {
-    "framework": "react",
-    "config": {
-      "SourceDir": "src",
-      "DistributionDir": "build",
-      "BuildCommand": "npm run-script build",
-      "StartCommand": "npm run-script start"
-    }
-  },
-  "providers": [
-    "awscloudformation"
-  ]
-}
