--- src/frontend/amplify/README.md
@@ -1,8 +1,0 @@
-# Getting Started with Amplify CLI
-This directory was generated by [Amplify CLI](https://docs.amplify.aws/cli).
-
-Helpful resources:
-- Amplify documentation: https://docs.amplify.aws.
-- Amplify CLI documentation: https://docs.amplify.aws/cli.
-- More details on this folder & generated files: https://docs.amplify.aws/cli/reference/files.
-- Join Amplify's community: https://amplify.aws/community/.
