--- src/frontend/amplify/backend/backend-config.json
@@ -1,1 +1,0 @@
-{}
