--- src/frontend/amplify/backend/types/amplify-dependent-resources-ref.d.ts
@@ -1,1 +1,0 @@
-export type AmplifyDependentResourcesAttributes = {}
