--- src/frontend/amplify/hooks/README.md
@@ -1,7 +1,0 @@
-# Command Hooks
-
-Command hooks can be used to run custom scripts upon Amplify CLI lifecycle events like pre-push, post-add-function, etc.
-
-To get started, add your script files based on the expected naming convention in this directory.
-
-Learn more about the script file naming convention, hook parameters, third party dependencies, and advanced configurations at https://docs.amplify.aws/cli/usage/command-hooks
