+++ src/frontend/src/proxy.js
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: MIT.
 
-const proxy_url = "http://localhost:5000"
+const proxy_url = "https://api.kovvuri.people.aws.dev";
 
-
-export default proxy_url
+export default proxy_url;
+
